requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
dataclasses-json==0.6.1
distro==1.8.0
exceptiongroup==1.1.3
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.4
google-cloud-aiplatform==1.36.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.13.0
google-cloud-vision==3.4.5
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.2
grpcio-status==1.59.2
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.333
langchain-experimental==0.0.39
langsmith==0.0.63
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==1.2.2
packaging==23.2
pandas==2.1.2
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vertexai==0.0.1
yarl==1.9.2
